[#1] 🔨Refactor  : member package 구조 변경
