+++ src/main/java/com/myinvestment/config/UserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package com.myinvestment.utils;
+package com.myinvestment.config;
 
 
 import com.myinvestment.dao.MemberDao;
