+++ src/main/java/com/myinvestment/service/LoginService.java
@@ -5,8 +5,8 @@ import com.myinvestment.dto.request.LoginRequestDto;
 import com.myinvestment.dto.request.MemberRequestDto;
 import com.myinvestment.dto.response.LoginResponseDto;
 import com.myinvestment.mapper.MemberMapper;
-import com.myinvestment.utils.DuplicateException;
-import com.myinvestment.utils.ErrorCode;
+import com.myinvestment.exception.DuplicateException;
+import com.myinvestment.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
